pyproject.toml
[tool.poetry]
name = "stone-desafio-qa-backend"
version = "0.1.0"
description = "Desafio QA Backend Stone"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
robotframework = "^7.0"
robotframework-requests = "^0.9.7"
robotframework-jsonlibrary = "^0.5"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
